+++ linux/flutter/generated_plugins.cmake
@@ -3,7 +3,10 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
+<<<<<<< Updated upstream
   file_selector_linux
+=======
+>>>>>>> Stashed changes
   url_launcher_linux
 )