+++ macos/Flutter/GeneratedPluginRegistrant.swift
@@ -6,13 +6,23 @@ import FlutterMacOS
 import Foundation
 
 import bonsoir_darwin
+<<<<<<< Updated upstream
 import file_selector_macos
+=======
+import file_picker
+import path_provider_foundation
+>>>>>>> Stashed changes
 import share_plus
 import shared_preferences_foundation
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   SwiftBonsoirPlugin.register(with: registry.registrar(forPlugin: "SwiftBonsoirPlugin"))
+<<<<<<< Updated upstream
   FileSelectorPlugin.register(with: registry.registrar(forPlugin: "FileSelectorPlugin"))
+=======
+  FilePickerPlugin.register(with: registry.registrar(forPlugin: "FilePickerPlugin"))
+  PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
+>>>>>>> Stashed changes
   SharePlusMacosPlugin.register(with: registry.registrar(forPlugin: "SharePlusMacosPlugin"))
   SharedPreferencesPlugin.register(with: registry.registrar(forPlugin: "SharedPreferencesPlugin"))
 }